backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
sqlalchemy[asyncio]==2.0.30
asyncpg==0.29.0
alembic==1.13.1
pydantic==2.7.1
pydantic-settings==2.2.1
celery==5.4.0
redis==5.0.4
aioredis==2.0.1
python-multipart==0.0.9
aiofiles==23.2.1
PyMuPDF
python-docx==1.1.0
Pillow==10.3.0
